+++ src/lib.rs
@@ -215,6 +215,8 @@
 //#[global_allocator]
 //static A: System = System;
 
+#![warn(missing_docs)]
+
 #[macro_use] extern crate log;
 #[macro_use] extern crate serde_derive;
 extern crate serde_json;